+++ _e2e/expect_test.ts
@@ -5,6 +5,8 @@ import {
   defineExpect,
   expect,
   objectContaining,
+  stringContaining,
+  stringMatching,
   test,
   toStartWith,
   toThrow,
@@ -39,4 +41,12 @@ test("objectContaining", () => {
   expect({ a: 1, b: "any" }).toEqual(
     objectContaining({ a: any(Number), b: any(String) }),
   );
+});
+
+test("should be match pattern", () => {
+  expect("hello! This is a good day.").toEqual(stringMatching(/good/));
+});
+
+test("should contain pattern", () => {
+  expect("hello! This is a good day.").toEqual(stringContaining("good"));
 });