+++ expect/mod.ts
@@ -11,6 +11,8 @@ import { arrayContaining } from "./array_containing.ts";
 import { objectContaining } from "./object_containing.ts";
 import { anyNumber } from "./any_number.ts";
 import { anyString } from "./any_string.ts";
+import { stringMatching } from "./string_matching.ts";
+import { stringContaining } from "./string_containing.ts";
 
 import type {
   AnyFn,
@@ -175,5 +177,7 @@ export {
   defineExpect,
   expect,
   objectContaining,
+  stringContaining,
+  stringMatching,
 };
 export type { Expected, MatcherMap };